+++ Application/app/controllers/students.controller.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const { json } = require('express');
-const connection = require('../../confing/connection.js')
+const connection = require('../../config/connection.js')
 const jwt = require('jsonwebtoken');
 
 const getStudents = (req, res) => {   
@@ -23,11 +23,16 @@ async function getOneStudent(req, res){
     const {document_num} = req.params
     jwt.verify(req.token, 'secretkey', async (error) =>{
         if(!error){
-            try{
-                const result = await connection.query(`select * from students where document_num = ${connection.escape(document_num)}`);
-                res.json(result);
-            }catch(error){
-                res.json(error);
+            let queryValidate = await connection.query(`SELECT * FROM students WHERE document_num = ${connection.escape(data.document_num)}`);
+            if(queryValidate.length === 1){
+                try{
+                    const result = await connection.query(`select * from students where document_num = ${connection.escape(document_num)}`);
+                    res.json(result);
+                }catch(error){
+                    res.json(error);
+                }
+            }else{
+                res.json({message: "El estudiante que busca, no existe."})
             }
         }else{
             res.sendStatus(403);
@@ -39,12 +44,17 @@ const addStudent = async (req, res) =>{
     const data = req.body;
     jwt.verify(req.token, 'secretkey', async (error) =>{
         if(!error){
-            try {
-                const result = await connection.query(`Insert into students(document_type, document_num, first_name, last_name, email, phone_number, semester, birthdate) values (${connection.escape(data.document_type)}, ${connection.escape(data.document_num)}, ${connection.escape(data.first_name)}, ${connection.escape(data.last_name)}, ${connection.escape(data.email)},${connection.escape(data.phone_number)}, ${connection.escape(data.semester)}, ${connection.escape(data.birthdate)})`)
-                res.json({message: `${connection.escape(data.first_name)} fue agregado correctamente.`})
-            } catch (error) {
-                res.json(error)
-            }
+            let queryValidate = await connection.query(`SELECT * FROM students WHERE document_num = ${connection.escape(data.document_num)}`);
+            if(queryValidate.length === 0){
+                try {
+                    const result = await connection.query(`Insert into students(document_type, document_num, first_name, last_name, email, phone_number, semester, birthdate) values (${connection.escape(data.document_type)}, ${connection.escape(data.document_num)}, ${connection.escape(data.first_name)}, ${connection.escape(data.last_name)}, ${connection.escape(data.email)},${connection.escape(data.phone_number)}, ${connection.escape(data.semester)}, ${connection.escape(data.birthdate)})`)
+                    res.json({message: `${connection.escape(data.first_name)} fue agregado correctamente.`})
+                } catch (error) {
+                    res.json(error)
+                }
+            }else{
+                res.json({message: "El estudiante que intenta ingresar ya existe."})
+            } 
         }else{
             res.sendStatus(403);
         }
@@ -56,11 +66,16 @@ const modifyStudent = async (req, res) =>{
     const {document_type, first_name, last_name, email, phone_number, semester, birthdate} = req.body; 
     jwt.verify(req.token, 'secretkey', async (error) =>{
         if(!error){
-            try {
-            const result = await connection.query(`update students set document_type = ${connection.escape(document_type)},first_name = ${connection.escape(first_name)}, last_name = ${connection.escape(last_name)}, email = ${connection.escape(email)}, phone_number = ${connection.escape(phone_number)}, semester = ${connection.escape(semester)}, birthdate = ${connection.escape(birthdate)} where document_num= ${connection.escape(document_num)}`);
-            res.json({message: `${connection.escape(first_name)} se ha modificado correctamente`})
-            } catch (error) {
-                res.json({message: `Aca ocurrio un error ${error}`});
+            let queryValidate = await connection.query(`SELECT * FROM students WHERE document_num = ${connection.escape(document_num)}`);
+            if(queryValidate.length === 1){
+                try {
+                    const result = await connection.query(`update students set document_type = ${connection.escape(document_type)},first_name = ${connection.escape(first_name)}, last_name = ${connection.escape(last_name)}, email = ${connection.escape(email)}, phone_number = ${connection.escape(phone_number)}, semester = ${connection.escape(semester)}, birthdate = ${connection.escape(birthdate)} where document_num= ${connection.escape(document_num)}`);
+                    res.json({message: `${connection.escape(first_name)} se ha modificado correctamente`})
+                    } catch (error) {
+                        res.json({message: `Ha ocurrido un error ${error}`});
+                    }
+            }else{
+                res.json({message: "EL usuario que intenta modificar no existe."})
             }
         }else{
             res.sendStatus(403);
@@ -72,11 +87,16 @@ const deleteStudent = async (req, res) => {
     const {document_num} = req.params;
     jwt.verify(req.token, 'secretkey', async (error) =>{
         if(!error){
-            try {
-                const result = await connection.query(`delete from students where document_num = ${connection.escape(document_num)}`)
-                res.json({message: "Estudiante eliminado correctamente"})
-            } catch (error) {
-                res.json(error);
+            let queryValidate = await connection.query(`SELECT * FROM students WHERE document_num = ${connection.escape(document_num)}`);
+            if(queryValidate.length === 1){
+                try {
+                    const result = await connection.query(`delete from students where document_num = ${connection.escape(document_num)}`)
+                    res.json({message: "Estudiante eliminado correctamente"})
+                } catch (error) {
+                    res.json(error);
+                }
+            }else{
+                res.json({message: "El estudiante que intenta eliminar no existe."})
             }
         }else{
             res.sendStatus(403);
